Removed switch cases and added codes text file
For thoughts of making it slight easier to control and maybe not just adding them everywhere. I removed all switch cases and made a codes text file. When it fetches the code, it will now go through the file and get the code. All codes in the text file are in code order.